refactor: renamed solc-references lib and extracted common function to lsp and path utils libs
