+++ libs/ast-extractor/Cargo.toml
@@ -2,7 +2,7 @@
 name = "osmium-libs-solidity-ast-extractor"
 description = "Extracts the AST from a Solidity file"
 license = "GPL-3.0-or-later"
-repository = "https://github.com/astrodevs-labs/osmium"
+repository = "https://github.com/astrodevs-labs/osmium-solidity"
 version = "0.1.3"
 edition = "2021"
