+++ libs/foundry-wrapper/src/utils/path.rs
@@ -13,15 +13,3 @@ pub fn find_projects_paths(root_path: &str) -> Result<Vec<PathBuf>, glob::Patter
         .map(|path| path.parent().unwrap().to_path_buf())
         .collect())
 }
-
-pub fn normalize_path(path: &str) -> String {
-    path.replace('\\', "/")
-        .replace("//", "/")
-        .replace("\\\\", "/")
-}
-
-pub fn join_path(base_path: &str, file: &str) -> String {
-    let mut path = PathBuf::from(base_path);
-    path.push(file);
-    path.to_str().unwrap().to_string().replace('\\', "/")
-}
