+++ libs/lsp-utils/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "osmium-libs-solidity-lsp-utils"
+version = "0.1.0"
+edition = "2021"
+authors = ["Astrodevs-Labs"]
+description = "Utils for lsp communication"
+license = "GPL-3.0-or-later"
+repository = "https://github.com/astrodevs-labs/osmium-solidity"
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tower-lsp = "0.20.0"
+osmium-libs-solidity-path-utils = { path = "../path-utils" }
