+++ libs/lsp-utils/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "osmium-libs-solidity-path-utils",
+  "license": "GNU-GPL-3.0",
+  "scripts": {
+    "build": "cargo build --release --target-dir ./target",
+    "format": "cargo fmt --all",
+    "format:check": "cargo fmt --check --all",
+    "lint": "cargo clippy --allow-dirty --allow-staged --all-targets --all-features --fix -- -D warnings",
+    "lint:check": "cargo clippy --all-targets --all-features -- -D warnings",
+    "test": "cargo test",
+    "publish": "cargo package --allow-dirty && cargo publish --allow-dirty",
+    "publish:check": "cargo package --allow-dirty && cargo publish --allow-dirty --dry-run"
+  }
+}
