+++ libs/references/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "solc-references"
+name = "osmium-libs-solidity-references"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,4 +8,5 @@ edition = "2021"
 [dependencies]
 solc-ast-rs-types = { version = "0.1.6", features = ["visit"]}
 osmium-libs-solidity-ast-extractor = { path="../ast-extractor" }
+osmium-libs-solidity-path-utils = { path = "../../libs/path-utils" }
 thiserror = "1.0.56"
