+++ libs/references/src/lib.rs
@@ -13,7 +13,7 @@ pub use solc_ast_rs_types::types::*;
 use types::InteractableNode;
 pub use types::{Location, Position};
 use usages::UsagesFinder;
-use utils::join_path;
+use osmium_libs_solidity_path_utils::join_path;
 
 use crate::utils::get_location;
