+++ servers/foundry-compiler-server/Cargo.toml
@@ -9,5 +9,7 @@ edition = "2021"
 
 [dependencies]
 osmium-libs-solidity-foundry-wrapper = { path = "../../libs/foundry-wrapper" }
+osmium-libs-solidity-path-utils = { path = "../../libs/path-utils" }
+osmium-libs-solidity-lsp-utils = { path = "../../libs/lsp-utils" }
 tower-lsp = "0.20.0"
 tokio = { version = "1.34.0", features = ["full"] }
