--- servers/foundry-compiler-server/Makefile
@@ -1,3 +1,0 @@
-build:
-	cargo build 
-	cp ../../target/debug/foundry-server ../../../extension/dist
