+++ servers/references-server/Cargo.toml
@@ -9,5 +9,6 @@ edition = "2021"
 [dependencies]
 tokio = { version = "1.36.0",  features = ["full"] }
 tower-lsp = "0.20.0"
-solc-references = { path = "../../libs/solc-references" }
+osmium-libs-solidity-references = { path = "../../libs/references" }
 glob = "0.3.1"
+osmium-libs-solidity-path-utils = { path = "../../libs/path-utils" }
