+++ servers/references-server/package.json
@@ -10,5 +10,4 @@
       "lint:check": "cargo clippy --all-targets --all-features -- -D warnings",  
       "test": "cargo test --workspace"
     }
-  }
-  
+}
