+++ servers/references-server/src/main.rs
@@ -2,13 +2,15 @@ mod utils;
 
 use crate::utils::*;
 
-use solc_references::*;
+use osmium_libs_solidity_references::*;
 use std::sync::Arc;
 use tokio::sync::Mutex;
 use tower_lsp::jsonrpc::Result;
 use tower_lsp::lsp_types::Location as LspLocation;
 use tower_lsp::lsp_types::*;
 use tower_lsp::{Client, LanguageServer, LspService, Server};
+
+use osmium_libs_solidity_path_utils::{escape_path, normalize_path};
 
 struct Backend {
     client: Client,
@@ -99,7 +101,7 @@ impl LanguageServer for Backend {
 
         let locations = self.references_provider.lock().await.get_references(
             &normalize_path(uri.path()),
-            solc_references::Position {
+            osmium_libs_solidity_references::Position {
                 line: position.line,
                 column: position.character,
             },
@@ -129,7 +131,7 @@ impl LanguageServer for Backend {
 
         let location = self.references_provider.lock().await.get_definition(
             &normalize_path(uri.path()),
-            solc_references::Position {
+            osmium_libs_solidity_references::Position {
                 line: position.line,
                 column: position.character,
             },
