+++ servers/references-server/src/utils.rs
@@ -2,33 +2,9 @@ use tower_lsp::lsp_types::{
     Location as LspLocation, Position as LspPosition, Range as LspRange, Url,
 };
 
-#[cfg(target_family = "windows")]
-pub fn normalize_path(path: &str) -> String {
-    let mut path = path.replace("%3A/", "://");
-    path.remove(0);
-    path.to_string()
-}
-
-#[cfg(not(target_family = "windows"))]
-pub fn normalize_path(path: &str) -> String {
-    path.to_string()
-}
-
-#[cfg(target_family = "windows")]
-pub fn escape_path(path: &str) -> String {
-    let mut path = path.replace("://", "%3A/");
-    path.insert(0, '/');
-    path.to_string()
-}
-
-#[cfg(not(target_family = "windows"))]
-pub fn escape_path(path: &str) -> String {
-    path.to_string()
-}
-
 pub fn location_to_lsp_location(
     new_uri: &Url,
-    location: &solc_references::Location,
+    location: &osmium_libs_solidity_references::Location,
 ) -> LspLocation {
     LspLocation {
         uri: new_uri.clone(),
