+++ servers/slither-server/Cargo.toml
@@ -18,3 +18,4 @@ thiserror = "1.0.50"
 glob = "0.3.1"
 toml = "0.8.8"
 tokio-util = "0.7.10"
+osmium-libs-solidity-path-utils = { path = "../../libs/path-utils" }
